requirements.txt
pywebview>=4.0
requests>=2.31.0
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.1.1
plyer>=2.1.0

# Phase 1 — Memory
chromadb>=0.4.0
networkx>=3.0
sentence-transformers>=2.2.0
numpy>=1.24.0
